Use provider API to configure tasks
Changes all tasks to use the provider API instead of plain fields. This
allows for late wiring, e.g. if one wants to set the value of
environmentVariables to a value that is computed by another task.
By doing this all tasks have been made abstract so Gradle will create an
implementation of the task and inject the property instance into it.